Moving save_encoding func to saver.py
